refactor: split apply guards and review planning into focused modules
